Add regression test for https://crbug.com/843523.
Test that document.querySelector(":target") works when called in the
window.load event.

This test is a subset of dom/nodes/Element-matches.html, the difference
is that this test is more lightweight. It causes the window.load event
to be sent earlier. There is a race condition. Currently, it will
only work when FrameLoader::ProcessFragment is called before
Document::ImplicitClose().

Probable fix:
https://chromium-review.googlesource.com/c/chromium/src/+/1057507

Bug: 843523
Change-Id: Ibf7b0d7621b18c3dfefaaa323bbd335d86ce0652